// src/lib/api-client/models/user.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DomainFlow API
 * DomainFlow API for domain generation, validation, and campaign management.
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: support@domainflow.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface User
 */
export interface User {
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'avatarUrl'?: string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'createdAt'?: string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'email'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
    'emailVerified'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'firstName'?: string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'id'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
    'isActive'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
    'isLocked'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'lastLoginAt'?: string;
    /**
     * 
     * @type {Array<object>}
     * @memberof User
     */
    'lastLoginIp'?: Array<object>;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'lastName'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
    'mfaEnabled'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'mfaLastUsedAt'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof User
     */
    'mustChangePassword'?: boolean;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof User
     */
    'updatedAt'?: string;
}